# add support of comments everywhere