--- a/pkg/astparser/parser_token_helpers.go
+++ b/pkg/astparser/parser_token_helpers.go
@@ -7,50 +7,22 @@ import (
 	"github.com/jensneuse/graphql-go-tools/pkg/lexer/token"
 )
 
-// hasNextToken - checks that we haven't reached eof
-func (p *Parser) hasNextToken() bool {
-	return p.currentToken+1 < p.maxTokens
-}
-
-// next - increments current token index if hasNextToken
-// otherwise returns current token
-func (p *Parser) next() int {
-	if p.hasNextToken() {
-		p.currentToken++
-	}
-	return p.currentToken
-}
-
 // read - increments currentToken index and return token if hasNextToken
 // otherwise returns keyword.EOF
 func (p *Parser) read() token.Token {
-	if p.hasNextToken() {
-		return p.tokens[p.next()]
-	}
-
-	return token.Token{
-		Keyword: keyword.EOF,
-	}
+	return p.tokenizer.Read()
 }
 
 // peek - returns token next to currentToken if hasNextToken
 // otherwise returns keyword.EOF
 func (p *Parser) peek() keyword.Keyword {
-	if p.hasNextToken() {
-		nextIndex := p.currentToken + 1
-		return p.tokens[nextIndex].Keyword
-	}
-	return keyword.EOF
+	return p.tokenizer.Peek()
 }
 
 // peekLiteral - returns token next to currentToken and token name as a ast.ByteSliceReference if hasNextToken
 // otherwise returns keyword.EOF
 func (p *Parser) peekLiteral() (keyword.Keyword, ast.ByteSliceReference) {
-	if p.hasNextToken() {
-		nextIndex := p.currentToken + 1
-		return p.tokens[nextIndex].Keyword, p.tokens[nextIndex].Literal
-	}
-	return keyword.EOF, ast.ByteSliceReference{}
+	return p.tokenizer.PeekLiteral()
 }
 
 // peekEquals - checks that next token is equal to key