Benchmark results:

On master:
```
goos: darwin
goarch: arm64
pkg: github.com/jensneuse/graphql-go-tools/pkg/astparser
BenchmarkParseStarwars
BenchmarkParseStarwars-8        	   22077	     47744 ns/op	 237.24 MB/s	    1210 B/op	      37 allocs/op
BenchmarkParseGithub
BenchmarkParseGithub-8          	     511	   2349871 ns/op	 250.10 MB/s	   84210 B/op	     971 allocs/op
BenchmarkSelectionSet
BenchmarkSelectionSet-8         	  991680	      1157 ns/op	 234.30 MB/s	       0 B/op	       0 allocs/op
BenchmarkIntrospectionQuery
BenchmarkIntrospectionQuery-8   	  205233	      5857 ns/op	 222.64 MB/s	       0 B/op	       0 allocs/op
BenchmarkKitchenSink
BenchmarkKitchenSink-8          	  247881	      4865 ns/op	 209.67 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse
BenchmarkParse-8                	  201710	      5800 ns/op	 230.16 MB/s	     160 B/op	       9 allocs/op
PASS
```

On this PR:
```
goos: darwin
goarch: arm64
pkg: github.com/jensneuse/graphql-go-tools/pkg/astparser
BenchmarkParseStarwars
BenchmarkParseStarwars-8        	   21216	     55375 ns/op	 204.55 MB/s	    1212 B/op	      37 allocs/op
BenchmarkParseGithub
BenchmarkParseGithub-8          	     450	   2583312 ns/op	 227.50 MB/s	   87275 B/op	     971 allocs/op
BenchmarkSelectionSet
BenchmarkSelectionSet-8         	  683020	      1617 ns/op	 167.62 MB/s	       0 B/op	       0 allocs/op
BenchmarkIntrospectionQuery
BenchmarkIntrospectionQuery-8   	  145411	      8040 ns/op	 162.20 MB/s	       0 B/op	       0 allocs/op
BenchmarkKitchenSink
BenchmarkKitchenSink-8          	  176130	      6686 ns/op	 152.56 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse
BenchmarkParse-8                	  158084	      7345 ns/op	 181.75 MB/s	     160 B/op	       9 allocs/op
PASS
```